+++ CHANGELOG.md
@@ -1,19 +1,32 @@
+<!-- markdownlint-disable MD024 no-duplicate-heading/no-duplicate-header -->
+
 # BuildMasterAutomation Changelog
 
 ## 2.0.1
+
+### Upgrade Instructions
+
+This is a continuation of the types of changes made in 2.0.0. See that release's changelog for upgrade instructions.
+
+* These functions now write an error if an item doesn't exist (for `Get`, `Remove`, and `Set` functions) or if an item
+already exists (for `New` functions). Add `-ErrorAction Ignore` to existing usages to preserve previous behavior:
+  * `Set-BMVariable`
+  * `Stop-BMRelease`
 
 ### Added
 
 #### Parameters
 
-* `Application` to the `Disable-BMApplication` function. This parameter accepts an application name, id, or
-application object and replaces the now-obsolete `ID` parameter.
+* `Application` to the `Disable-BMApplication` function. This parameter accepts an application name, id, or application
+object and replaces the now-obsolete `ID` parameter.
 * `Set-BMVariable`:
   * `Application`: accepts application ids, names, or application objects.
   * `ApplicationGroup`: accepts application group ids, names, or application group objects.
   * `Environment`: accepts environment ids, names, or environment objects.
   * `Server`: accepts server ids, names, or server objects.
   * `ServerRole`: accepts server role ids, names, or server names.
+* `Application` to the `Stop-BMRelease` function. This parameter accepts an application name, id, or application object
+and replaces the now-obsolete `ID` parameter.
 
 ### Deprecated
 
@@ -24,6 +37,7 @@ application object and replaces the now-obsolete `ID` parameter.
   * `EnvironmentName`; use `Environment` instead.
   * `ServerName`; use `Server` instead.
   * `ServerRoleName`; use `ServerRole` instead.
+* `ApplicationID` on the `Stop-BMRelease` function. Use the new `Application` parameter isntead.
 
 ## 2.0.0
 
@@ -62,6 +76,7 @@ already exists (for `New` functions). Add `-ErrorAction Ignore` to existing usag
   * `Remove-BMVariable`
   * `Set-BMRelease`
   * `Set-BMVariable`
+  * `Stop-BMRelease`
 * Remove usages of the `Get-BMDeployment` function's `Build`, `Release`, and `Application` parameters. The BuildMaster
 [Release and build deployment API](https://docs.inedo.com/docs/buildmaster-reference-api-release-and-build) no longer
 supports getting deploys for builds, releases, and applications.